backend/backend/requirements.txt
boto3==1.24.84
celery == 5.2.6
colorlog == 6.6.0
django==4.1.3
django-admin-menu==1.6
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==4.8.0
django-elasticsearch-dsl==7.2.2
drf-spectacular==0.24.2
django-storages==1.13.1
djoser==2.1.0
drf-yasg==1.21.3
elasticsearch==7
elasticsearch-dsl==7.4.0
gunicorn == 20.1.0
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.3
redis==4.2.2
requests==2.28.1
social-auth-app-django==4.0.0
whitenoise==6.2.0
watchfiles==0.18.1
